Remove AbstractPoWBlockstore, getBlockHashByNumber moved to AionBlockStore
